+++ test/support/adapter.ex
@@ -42,6 +42,21 @@ defmodule Spandex.TestAdapter do
     end
   end
 
+  @doc """
+  Injects test HTTP headers to represent the specified SpanContext
+  """
+  @impl Spandex.Adapter
+  @spec inject_context(Spandex.headers(), SpanContext.t(), Tracer.opts()) :: Spandex.headers()
+  def inject_context(headers, %SpanContext{trace_id: trace_id, parent_id: parent_id, priority: priority}, _opts) do
+    [
+      {"x-test-trace-id", to_string(trace_id)},
+      {"x-test-parent-id", to_string(parent_id)},
+      {"x-test-sampling-priority", to_string(priority)}
+    ] ++ headers
+  end
+
+  # Private Helpers
+
   @spec get_first_header(conn :: Plug.Conn.t(), header_name :: binary) :: binary | nil
   defp get_first_header(conn, header_name) do
     conn